Validation Back End

+++ Organiser.Core/Organiser.Core/CQRS/Resources/Tasks/Tasks/Handlers/DeleteTaskCommandHandler.cs
@@ -17,6 +17,16 @@ namespace Organiser.Core.CQRS.Resources.Tasks.Tasks.Handlers
             if (task == null)
                 throw new TaskNotFoundException("Nie znaleziono zadania");
 
+            var taskNotesCount = context.TasksNotes.Count(x => x.TNTGID == task.TGID);
+
+            if (taskNotesCount > 0)
+                throw new TaskConatinsTaskNotesException("Do zadania przypisane są notatki! Usuń je najpierw!");
+
+            var taskSubTasksCount = context.TasksSubTasks.Count(x => x.TSTTGID == task.TGID);
+
+            if (taskSubTasksCount > 0)
+                throw new TaskConatinsSubTasksException("Do zadania przypisane są podzadania! Usuń je najpierw!");
+
             context.DeleteTask(task);
             context.SaveChanges();
         }
